Update README.md
"How to Train Your Dragon" is a delightful animated film that takes us to the Viking village of Berk, where battling dragons is a way of life. The story follows Hiccup, a young Viking who is more inventive than brawny, struggling to gain the approval of his father, Stoick the Vast, and fit into the village’s dragon-fighting tradition. Everything changes when Hiccup encounters a mysterious and rare Night Fury dragon that he names Toothless. Instead of fighting Toothless, Hiccup befriends and trains him, discovering that dragons are not the vicious creatures the Vikings believed them to be.

Through their friendship, Hiccup and Toothless challenge the long-standing prejudices of the village and eventually uncover the true reason behind the dragons' attacks. The film beautifully blends action, adventure, and heartfelt moments, highlighting themes of friendship, understanding, and challenging the status quo.